various frontend bug fixes

+++ frontend/src/Components/QandA.tsx
@@ -43,7 +43,7 @@ interface State {
   questions: Question[];
   sortBy: string;
   loading: boolean;
-  user: User;
+  user: User | null;
   writingQuestion: boolean;
   answeringQuestion: Question | null;
 }
@@ -130,6 +130,9 @@ export default class QandA extends Component<Props, State> {
   };
 
   onUpvoteQuestionClicked = (question: Question) => {
+    if (!this.state.user) {
+      return;
+    }
     question.downvoters.includes(this.state.user.username)
       ? QandAService.removeQuestionDownvote(
           this.state.user.id,
@@ -159,6 +162,9 @@ export default class QandA extends Component<Props, State> {
   };
 
   onDownVoteQuestionClicked = (question: Question) => {
+    if (!this.state.user) {
+      return;
+    }
     question.upvoters.includes(this.state.user.username)
       ? QandAService.removeQuestionUpvote(
           this.state.user.id,
@@ -185,6 +191,9 @@ export default class QandA extends Component<Props, State> {
   };
 
   onUpvoteAnswerClicked = (answer: Answer) => {
+    if (!this.state.user) {
+      return;
+    }
     answer.downvoters.includes(this.state.user.username)
       ? QandAService.removeAnswerDownvote(
           this.state.user.id,
@@ -211,6 +220,9 @@ export default class QandA extends Component<Props, State> {
   };
 
   onDownVoteAnswerClicked = (answer: Answer) => {
+    if (!this.state.user) {
+      return;
+    }
     answer.upvoters.includes(this.state.user.username)
       ? QandAService.removeAnswerUpvote(
           this.state.user.id,
@@ -237,6 +249,9 @@ export default class QandA extends Component<Props, State> {
   };
 
   onSubmitClicked = (content: string) => {
+    if (!this.state.user) {
+      return;
+    }
     const params = this.props.streamId
       ? {
           userId: this.state.user.id,
@@ -331,27 +346,29 @@ export default class QandA extends Component<Props, State> {
       <Box>
         <Box direction="row" justify="between" align="center" id="question">
           <Box direction="row" gap="small">
-            <Box
-              height="100%"
-              justify="between"
-              align="center"
-              pad="none"
-              margin="none"
-            >
-              <CaretUpOutlined
-                style={{
-                  fontSize: 35,
-                  marginTop: -5,
-                  marginBottom: 5,
-                  color: upvoteColor,
-                }}
-                onClick={() => this.onUpvoteQuestionClicked(question)}
-              />
-              <CaretDownOutlined
-                style={{ fontSize: 35, color: downvoteColor }}
-                onClick={() => this.onDownVoteQuestionClicked(question)}
-              />
-            </Box>
+            {this.state.user && (
+              <Box
+                height="100%"
+                justify="between"
+                align="center"
+                pad="none"
+                margin="none"
+              >
+                <CaretUpOutlined
+                  style={{
+                    fontSize: 35,
+                    marginTop: -5,
+                    marginBottom: 5,
+                    color: upvoteColor,
+                  }}
+                  onClick={() => this.onUpvoteQuestionClicked(question)}
+                />
+                <CaretDownOutlined
+                  style={{ fontSize: 35, color: downvoteColor }}
+                  onClick={() => this.onDownVoteQuestionClicked(question)}
+                />
+              </Box>
+            )}
             <Box>
               <Box direction="row" gap="small">
                 <Box align="center" height="100%" justify="between">
@@ -380,24 +397,26 @@ export default class QandA extends Component<Props, State> {
               <Box>{this.renderQuestionOrAnswerContent(question.content)}</Box>
             </Box>
           </Box>
-          <Box
-            className="hides"
-            direction="row"
-            width="100px"
-            height="40px"
-            round="small"
-            background="#61EC9F"
-            margin="none"
-            justify="center"
-            align="center"
-            gap="small"
-            onClick={() => this.onAnswerClicked(question)}
-            focusIndicator={false}
-          >
-            <Text color="white" weight="bold">
-              Answer
-            </Text>
-          </Box>
+          {this.state.user && (
+            <Box
+              className="hides"
+              direction="row"
+              width="100px"
+              height="40px"
+              round="small"
+              background="#61EC9F"
+              margin="none"
+              justify="center"
+              align="center"
+              gap="small"
+              onClick={() => this.onAnswerClicked(question)}
+              focusIndicator={false}
+            >
+              <Text color="white" weight="bold">
+                Answer
+              </Text>
+            </Box>
+          )}
         </Box>
         <Box
           margin={{ left: "xlarge", top: "small" }}
@@ -421,27 +440,29 @@ export default class QandA extends Component<Props, State> {
         : "black";
     return (
       <Box direction="row" gap="small">
-        <Box
-          height="100%"
-          justify="between"
-          align="center"
-          pad="none"
-          margin="none"
-        >
-          <CaretUpOutlined
-            style={{
-              fontSize: 35,
-              marginTop: -5,
-              marginBottom: 5,
-              color: upvoteColor,
-            }}
-            onClick={() => this.onUpvoteAnswerClicked(answer)}
-          />
-          <CaretDownOutlined
-            style={{ fontSize: 35, color: downvoteColor }}
-            onClick={() => this.onDownVoteAnswerClicked(answer)}
-          />
-        </Box>
+        {this.state.user && (
+          <Box
+            height="100%"
+            justify="between"
+            align="center"
+            pad="none"
+            margin="none"
+          >
+            <CaretUpOutlined
+              style={{
+                fontSize: 35,
+                marginTop: -5,
+                marginBottom: 5,
+                color: upvoteColor,
+              }}
+              onClick={() => this.onUpvoteAnswerClicked(answer)}
+            />
+            <CaretDownOutlined
+              style={{ fontSize: 35, color: downvoteColor }}
+              onClick={() => this.onDownVoteAnswerClicked(answer)}
+            />
+          </Box>
+        )}
         <Box>
           <Box direction="row" gap="small">
             <Box align="center" height="100%" justify="between">
@@ -497,7 +518,7 @@ export default class QandA extends Component<Props, State> {
               New
             </Text>
           </Box>
-          {!this.props.streamer && (
+          {!this.props.streamer && this.state.user && (
             <Box
               direction="row"
               width="100px"
